@@ -3,36 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/data-grid": "^4.0.0-alpha.37",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@stripe/react-stripe-js": "^1.4.1",
-    "@stripe/stripe-js": "^1.17.1",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.1",
-    "chart.js": "^3.5.1",
-    "country-state-city": "^3.0.1",
-    "overlay-navbar": "^1.0.4",
-    "react": "^17.0.2",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "axios": "^0.27.2",
+    "axis": "^1.0.0",
+    "overlay-navbar": "^1.2.3",
+    "react": "^18.1.0",
     "react-alert": "^7.0.3",
     "react-alert-template-basic": "^1.0.2",
-    "react-chartjs-2": "^3.0.4",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
     "react-helmet": "^6.1.0",
-    "react-icons": "^4.2.0",
-    "react-js-pagination": "^3.0.3",
-    "react-material-ui-carousel": "^2.3.1",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.1.1",
+    "react-icons": "^4.4.0",
+    "react-rating-stars-component": "^2.2.0",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
     "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.3.0",
-    "web-vitals": "^1.1.2",
-    "webfontloader": "^1.6.28"
+    "redux-thunk": "^2.4.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -57,6 +47,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "proxy": "http://192.168.29.21:4000"
+  }
 }
